+++ server/src/tests/cases/compiler/super2.c
@@ -1,0 +1,5 @@
+// super accessor for an efun
+// this is allowed even though this file doesn't inherit anything
+void write() {
+    efun::write("");
+}