+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## 1.1.8
 
--   Fix: [Symbol is resolving to efun instead of sefun #95](<https://github.com/jlchmura/lpc-language-server/issues/95>)
+-   Fix: [Symbol is resolving to efun instead of sefun #95](https://github.com/jlchmura/lpc-language-server/issues/95)
+-   Fix: [Symbols with super prefix do not resolve #97](https://github.com/jlchmura/lpc-language-server/issues/97)
 -   Added ability to use `@callback` and `@template` tags in function signatures.
+-   Enhance efun definitions
 
 ## 1.1.7